Move sources to src/ and rename
