+++ src/Detector.h
