+++ src/OpenScanDevice.h
