+++ src/OpenScanSettings.c
@@ -1,7 +1,7 @@
-#include "NIDAQ_OpenScanSettings.h"
+#include "OpenScanSettings.h"
 
-#include "NIDAQ_DeviceImplData.h"
-#include "OScNIDAQ.h"
+#include "DAQConfig.h"
+#include "DeviceImplData.h"
 
 #include <OpenScanDeviceLib.h>
