+++ src/meson.build
@@ -1,0 +1,13 @@
+openscan_nidaq_src = files(
+    'Acquisition.c',
+    'Clock.c',
+    'DAQConfig.c',
+    'Detector.c',
+    'DeviceImplData.c',
+    'OpenScanDevice.c',
+    'OpenScanModule.c',
+    'OpenScanSettings.c',
+    'ParkUnpark.c',
+    'Scanner.c',
+    'Waveform.c',
+)
